# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FITS Backend API
  version: 1.0.0
  description: |
    # FITS (Flexible IT Training System) Backend API

    Production-ready REST API for managing students, teachers, authentication, and signing workflows.

    ## Features
    - JWT-based authentication with refresh tokens
    - Role-based access control (Admin, Teacher, Student)
    - Invitation system for user onboarding
    - Student and teacher management with pagination
    - Digital signing workflow for reports
    - Rate limiting and security headers

    ## Authentication Flow
    1. **Bootstrap Admin**: `POST /api/v1/bootstrap/init` (one-time setup)
    2. **Login**: `POST /api/v1/auth/login` (get access + refresh tokens)
    3. **Use Access Token**: Include `Authorization: Bearer {access_token}` header
    4. **Refresh Token**: `POST /api/v1/auth/refresh` when access token expires
    5. **Logout**: `POST /api/v1/auth/logout` to invalidate tokens

    ## Rate Limiting
    - Global: 100 requests/minute per IP
    - Per-user: Varies by role (Admin: 200/min, Teacher: 100/min, Student: 50/min)

    ## Pagination
    All list endpoints support:
    - `page` (default: 1, min: 1)
    - `limit` (default: 20, max: 100)
  termsOfService: https://fits.example.com/terms
  contact:
    name: API Support Team
    email: support@fits.example.com
    url: https://fits.example.com/support
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/student:
    get:
      summary: List students with pagination
      description: |
        Retrieve a paginated list of students.
        Supports filtering via query parameters.
      operationId: listStudents
      tags: [Students]
      security: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of students
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedStudentsResponse'
              examples:
                success:
                  summary: First page with 2 students
                  value:
                    success: true
                    page: 1
                    limit: 20
                    total_count: 150
                    total_pages: 8
                    data:
                      - uuid: 550e8400-e29b-41d4-a716-446655440000
                        email: max.mustermann@example.com
                        first_name: Max
                        last_name: Mustermann
                        teacher_id: 550e8400-e29b-41d4-a716-446655440010
                        created_at: "2025-09-30T12:00:00Z"
                        updated_at: "2025-09-30T12:00:00Z"
                      - uuid: 550e8400-e29b-41d4-a716-446655440001
                        email: lisa.mueller@example.com
                        first_name: Lisa
                        last_name: Mueller
                        teacher_id: 550e8400-e29b-41d4-a716-446655440010
                        created_at: "2025-09-29T10:30:00Z"
                        updated_at: "2025-09-29T10:30:00Z"
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new student
      description: |
        Create a new student record. Requires admin role.
        Email must be unique.
      operationId: createStudent
      tags: [Students]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentRequest'
            examples:
              withTeacher:
                summary: Student with assigned teacher
                value:
                  email: max.mustermann@example.com
                  first_name: Max
                  last_name: Mustermann
                  teacher_id: 550e8400-e29b-41d4-a716-446655440010
              withoutTeacher:
                summary: Student without teacher (assign later)
                value:
                  email: lisa.mueller@example.com
                  first_name: Lisa
                  last_name: Mueller
      responses:
        '201':
          description: Student created successfully
          headers:
            Location:
              schema:
                type: string
              description: URI of created student
              example: /api/v1/student/550e8400-e29b-41d4-a716-446655440000
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Student'
              examples:
                success:
                  value:
                    success: true
                    message: Student created successfully
                    data:
                      uuid: 550e8400-e29b-41d4-a716-446655440000
                      email: max.mustermann@example.com
                      first_name: Max
                      last_name: Mustermann
                      teacher_id: 550e8400-e29b-41d4-a716-446655440010
                      created_at: "2025-10-22T14:30:00Z"
                      updated_at: "2025-10-22T14:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emailConflict:
                  value:
                    success: false
                    code: 409
                    error: Conflict
                    details: A student with this email already exists
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/student/{uuid}:
    get:
      summary: Get student by UUID
      description: Retrieve detailed information about a specific student
      operationId: getStudent
      tags: [Students]
      security: []
      parameters:
        - $ref: '#/components/parameters/UuidParam'
      responses:
        '200':
          description: Student found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Student'
              examples:
                success:
                  value:
                    success: true
                    data:
                      uuid: 550e8400-e29b-41d4-a716-446655440000
                      email: max.mustermann@example.com
                      first_name: Max
                      last_name: Mustermann
                      teacher_id: 550e8400-e29b-41d4-a716-446655440010
                      created_at: "2025-09-30T12:00:00Z"
                      updated_at: "2025-09-30T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update student information
      description: |
        Update an existing student's information. Requires admin role.
        Supports partial updates - only provided fields will be updated.
      operationId: updateStudent
      tags: [Students]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UuidParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStudentRequest'
            examples:
              updateEmail:
                summary: Update email only
                value:
                  email: max.new@example.com
              updateAll:
                summary: Update all fields
                value:
                  email: moritz.schmidt@example.com
                  first_name: Moritz
                  last_name: Schmidt
                  teacher_id: 550e8400-e29b-41d4-a716-446655440011
      responses:
        '200':
          description: Student updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Student'
              examples:
                success:
                  value:
                    success: true
                    message: Student updated successfully
                    data:
                      uuid: 550e8400-e29b-41d4-a716-446655440000
                      email: moritz.schmidt@example.com
                      first_name: Moritz
                      last_name: Schmidt
                      teacher_id: 550e8400-e29b-41d4-a716-446655440011
                      created_at: "2025-09-30T12:00:00Z"
                      updated_at: "2025-10-22T14:35:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emailConflict:
                  value:
                    success: false
                    code: 409
                    error: Conflict
                    details: Email already in use by another student
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete a student
      description: |
        Soft delete a student from the system. Requires admin role.
        Student data is marked as deleted but not physically removed.
      operationId: deleteStudent
      tags: [Students]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UuidParam'
      responses:
        '204':
          description: Student deleted successfully (no content)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== TEACHERS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.

        **Format**: `Authorization: Bearer {access_token}`

        **How to get tokens:**
        1. Login: `POST /api/v1/auth/login`
        2. Use returned `access_token` in Authorization header
        3. Refresh when expired: `POST /api/v1/auth/refresh`
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitParam:
      name: limit
      in: query
      required: false
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    UuidParam:
      name: uuid
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidJson:
              summary: Invalid JSON
              value:
                success: false
                code: 400
                error: Bad Request
                details: Invalid JSON format
            invalidUuid:
              summary: Invalid UUID format
              value:
                success: false
                code: 400
                error: Bad Request
                details: Invalid UUID format
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: Missing authorization header
              value:
                success: false
                code: 401
                error: Unauthorized
                details: Missing authorization token
            invalidToken:
              summary: Invalid or expired token
              value:
                success: false
                code: 401
                error: Unauthorized
                details: Invalid or expired token
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            insufficientPermissions:
              summary: Admin role required
              value:
                success: false
                code: 403
                error: Forbidden
                details: Admin role required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            studentNotFound:
              summary: Student not found
              value:
                success: false
                code: 404
                error: Not Found
                details: Student not found
            teacherNotFound:
              summary: Teacher not found
              value:
                success: false
                code: 404
                error: Not Found
                details: Teacher not found
    ValidationError:
      description: Validation error - invalid field values
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            passwordTooShort:
              summary: Password validation failed
              value:
                success: false
                code: 422
                error: Validation Error
                details: Password must be at least 8 characters long
            invalidEmail:
              summary: Email validation failed
              value:
                success: false
                code: 422
                error: Validation Error
                details: Invalid email format
            missingField:
              summary: Required field missing
              value:
                success: false
                code: 422
                error: Validation Error
                details: "first_name is required"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            serverError:
              summary: Generic server error
              value:
                success: false
                code: 500
                error: Internal Server Error
                details: An unexpected error occurred. Please try again later.
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
        data:
          type: object
          nullable: true
          description: Response payload (varies by endpoint)
    ErrorResponse:
      type: object
      required: [success, code, error]
      properties:
        success:
          type: boolean
          example: false
        code:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type/category
          example: Bad Request
        details:
          type: string
          description: Human-readable error details
          example: Invalid email format
    PaginationMeta:
      type: object
      required: [success, page, limit, total_count, total_pages]
      properties:
        success:
          type: boolean
          example: true
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Items per page
          example: 20
        total_count:
          type: integer
          description: Total number of items
          example: 150
        total_pages:
          type: integer
          description: Total number of pages
          example: 8

    # ==================== BOOTSTRAP ====================
...
    Student:
      type: object
      required: [uuid, email, first_name, last_name, created_at, updated_at]
      properties:
        uuid:
          type: string
          format: uuid
          description: Student's unique identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          description: Student's email address
          example: max.mustermann@example.com
        first_name:
          type: string
          minLength: 1
          maxLength: 100
          description: Student's first name
          example: Max
        last_name:
          type: string
          minLength: 1
          maxLength: 100
          description: Student's last name
          example: Mustermann
        teacher_id:
          type: string
          format: uuid
          nullable: true
          description: UUID of assigned teacher (optional)
          example: 550e8400-e29b-41d4-a716-446655440010
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2025-09-30T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-09-30T12:00:00Z"
    CreateStudentRequest:
      type: object
      required: [email, first_name, last_name]
      properties:
        email:
          type: string
          format: email
          description: Student's email address (must be unique)
          example: max.mustermann@example.com
        first_name:
          type: string
          minLength: 1
          maxLength: 100
          description: Student's first name
          example: Max
        last_name:
          type: string
          minLength: 1
          maxLength: 100
          description: Student's last name
          example: Mustermann
        teacher_id:
          type: string
          format: uuid
          nullable: true
          description: UUID of assigned teacher (optional - can be assigned later)
          example: 550e8400-e29b-41d4-a716-446655440010
    UpdateStudentRequest:
      type: object
      description: All fields are optional - only provided fields will be updated
      properties:
        email:
          type: string
          format: email
          description: New email address
          example: max.new@example.com
        first_name:
          type: string
          minLength: 1
          maxLength: 100
          description: New first name
          example: Moritz
        last_name:
          type: string
          minLength: 1
          maxLength: 100
          description: New last name
          example: Schmidt
        teacher_id:
          type: string
          format: uuid
          nullable: true
          description: New teacher UUID (use null to unassign)
          example: 550e8400-e29b-41d4-a716-446655440011
    PaginatedStudentsResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationMeta'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Student'

    # ==================== TEACHERS ====================
...